Remove unused code from lexer.
Removes cue-setting keywords, "NOTE" keyword, integers/percentages, timestamps,
and operators like "-->" and "." from the lexer code, as these are not used in
the code.

+++ src/libwebvtt/cue.c
@@ -494,10 +494,10 @@ webvtt_cue_validate_set_settings( webvtt_parser self, webvtt_cue *cue,
   if( !cue || !settings ) {
     return WEBVTT_INVALID_PARAM;
   }
-  length = webvtt_string_length( settings );
+  length = (int)webvtt_string_length( settings );
   if( ( eol = strchr( webvtt_string_text( settings ), '\r' ) )
       || ( eol = strchr( webvtt_string_text( settings ), '\n' ) ) ) {
-    length = ( eol - webvtt_string_text( settings ) );
+    length = (int)( eol - webvtt_string_text( settings ) );
   }
 
   if( self ) {
